separation of device.ts into different files by type
