+++ src/types/devices/index.ts
@@ -1,0 +1,6 @@
+// src/devices/index.ts
+
+export { Device } from "./device";
+export { Router } from "./router";
+export { Server } from "./server";
+export { Pc } from "./pc";
