+++ src/types/edge.ts
@@ -1,6 +1,6 @@
 import { Graphics } from "pixi.js";
 import { ViewGraph } from "./graphs/viewgraph";
-import { Device } from "./device";
+import { Device } from "./devices/index"; // Import the Device class
 import { deselectElement, selectElement } from "./viewportManager";
 import { RightBar } from "..";
