+++ src/types/viewportManager.ts
@@ -1,6 +1,6 @@
 import { GlobalContext } from "./../index";
 import { DataGraph, GraphNode } from "./graphs/datagraph";
-import { Device, Pc, Router, Server } from "./device";
+import { Device, Pc, Router, Server } from "./devices/index";
 import { Edge } from "./edge";
 import { RightBar } from "../index"; // Ensure the path is correct
